Move Render into Frontend
Frontend already exports these functions through
wrappers so this adds 10 lines and removes one
module.
